requirements.txt
transformers>=4.30.0
torch>=2.0.0
streamlit>=1.28.0
PyPDF2>=3.0.0
python-docx>=0.8.11
google-generativeai>=0.3.0
python-dotenv>=1.0.0
